output/requirements.txt
Flask>=3.0
Flask-CORS>=4.0
Flask-JWT-Extended>=4.6
Flask-SQLAlchemy>=3.0
SQLAlchemy>=2.0
Werkzeug>=3.0
passlib>=1.7
python-dotenv>=1.0
marshmallow>=3.0
marshmallow-sqlalchemy>=1.0
python-dateutil>=2.8
bcrypt>=4.0